Revert RandoEvent file segmentation
