+++ src/engine/behavior_script.c
@@ -1,5 +1,5 @@
 #include <libultraship.h>
-#include "port/hooks/list/RandoEvent.h"
+#include "port/hooks/list/PlayerEvent.h"
 
 #include "sm64.h"
 #include "behavior_data.h"
