+++ src/game/macro_special_objects.c
@@ -11,7 +11,7 @@
 
 #include "special_presets.h"
 
-#include "port/hooks/list/RandoEvent.h"
+#include "port/hooks/list/PlayerEvent.h"
 #include "port/mods/PortEnhancements.h"
 
 /*
